misc: ui.table clean code
