+++ src/js/components/layout/ui.table/utils.js
@@ -58,6 +58,77 @@ _.convertBeginEnd = function(str) {
         str = str.replace(/}/g, END);
     }
     return str;
+};
+
+var hasChildren = function(root) {
+    return root.children && root.children.length > 0
+};
+
+_.getHeaders = function(columns) {
+    var headers = [];
+    var extractHeaders = function(columns, depth) {
+        columns.forEach(function(column) {
+            if(hasChildren(column)) {
+                column._dataColumn = extractHeaders(column.children, depth + 1);
+            }
+            if(!headers[depth]) {
+                headers[depth] = [];
+            }
+            // 计算深度和宽度
+            if(hasChildren(column)) {
+                column.childrenDepth = 1 + column.children.reduce(function(previous, current) {
+                    return current.childrenDepth > previous ? current.childrenDepth : previous;
+                }, -1);
+                column.headerColSpan = column.children.reduce(function(previous, current) {
+                    return previous + (current.headerColSpan || 0);
+                }, 0);
+            } else {
+                column.childrenDepth = 0;
+                column.headerColSpan = 1;
+            }
+            headers[depth].push(column);
+        });
+    };
+    extractHeaders(columns, 0);
+    return headers;
+};
+
+_.getLeaves = function(tree) {
+    var res = [];
+    var extractLeaves = function(root) {
+        if(root.forEach) {
+            return root.forEach(function(item) {
+                if(hasChildren(item)) {
+                    extractLeaves(item.children);
+                } else {
+                    res.push(item);
+                }
+            });
+        }
+    };
+    extractLeaves(tree);
+    return res;
+};
+
+var getNum = _.getNum = function(str) {
+    return +((str+'').split('px')[0]);
+};
+
+_.setElementValue = function(ele, prop, val) {
+    if(ele) {
+        ele[prop] = val;
+    }
+};
+
+_.getElementHeight = function(ele) {
+    var computedStyle = window.getComputedStyle(ele);
+    var height = getNum(computedStyle.marginTop)
+            + getNum(computedStyle.borderTopWidth)
+            + getNum(ele.offsetHeight)
+            + getNum(computedStyle.borderBottomWidth)
+            + getNum(computedStyle.marginBottom);
+    return height;
 }
+
 
 module.exports = _;
